Export bugsnag hook
